requirements.txt
pre-commit
einops==0.8.0
Flask==3.0.3
imageio==2.35.1
matplotlib==3.9.2
mediapipe==0.10.14
motpy==0.0.10
numpy==1.24.3
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
pandas==2.2.3
PyYAML==6.0.1
rootutils==1.0.7
scikit_learn==1.4.2
scipy==1.14.1
torch==2.3.0
torchmetrics==1.4.0.post0
tqdm==4.66.4
